Removed Old WinForms Code

+++ Tools/SelectTool.cs
@@ -20,11 +20,11 @@ namespace AAP
 
             artMatrixPosition = new(Math.Clamp(artMatrixPosition.X, 0, App.CurrentArt.Width), Math.Clamp(artMatrixPosition.Y, 0, App.CurrentArt.Height));
 
-            int startX = Math.Clamp(artMatrixPosition.X > StartPoint.X ? StartPoint.X : artMatrixPosition.X, 0, App.CurrentArt.Width);
-            int startY = Math.Clamp(artMatrixPosition.Y > StartPoint.Y ? StartPoint.Y : artMatrixPosition.Y, 0, App.CurrentArt.Height);
+            int startX = (int)Math.Clamp(artMatrixPosition.X > StartPoint.X ? StartPoint.X : artMatrixPosition.X, 0, App.CurrentArt.Width);
+            int startY = (int)Math.Clamp(artMatrixPosition.Y > StartPoint.Y ? StartPoint.Y : artMatrixPosition.Y, 0, App.CurrentArt.Height);
 
-            int sizeX = Math.Clamp(artMatrixPosition.X > StartPoint.X ? artMatrixPosition.X - StartPoint.X + 1 : StartPoint.X - artMatrixPosition.X + 1, 0, App.CurrentArt.Width - Math.Min(StartPoint.X, artMatrixPosition.X));
-            int sizeY = Math.Clamp(artMatrixPosition.Y > StartPoint.Y ? artMatrixPosition.Y - StartPoint.Y + 1 : StartPoint.Y - artMatrixPosition.Y + 1, 0, App.CurrentArt.Height - Math.Min(StartPoint.Y, artMatrixPosition.Y));
+            int sizeX = (int)Math.Clamp(artMatrixPosition.X > StartPoint.X ? artMatrixPosition.X - StartPoint.X + 1 : StartPoint.X - artMatrixPosition.X + 1, 0, App.CurrentArt.Width - Math.Min(StartPoint.X, artMatrixPosition.X));
+            int sizeY = (int)Math.Clamp(artMatrixPosition.Y > StartPoint.Y ? artMatrixPosition.Y - StartPoint.Y + 1 : StartPoint.Y - artMatrixPosition.Y + 1, 0, App.CurrentArt.Height - Math.Min(StartPoint.Y, artMatrixPosition.Y));
 
             App.Selected = new(startX, startY, sizeX, sizeY);
         }
